refactor(sets): rename ArtistsTab folder to SetsTab
Aligns the directory name with the sets route it backs, per review
feedback on chiptus/UpLine#279.
