+++ src/pages/EditionView/tabs/SetsTab/MultiArtistSetCard.tsx
