+++ src/pages/EditionView/tabs/SetsTab/SetCard/MultiArtistInfo.tsx
