+++ src/pages/EditionView/tabs/SetsTab/SetCard/SetVotingButtons.tsx
