+++ src/pages/EditionView/tabs/SetsTab/SetCard/SocialPlatformLink.tsx
