+++ src/pages/EditionView/tabs/SetsTab/SetCard/SocialPlatformLinkList.tsx
