+++ src/pages/EditionView/tabs/SetsTab/SetCard/SocialPlatformUtils.tsx
