+++ src/pages/EditionView/tabs/SetsTab/SetCard/social-platform-logos.json
