+++ src/pages/EditionView/tabs/SetsTab/SingleArtistSetCard.tsx
